ci: switch to rust specific release actions
Use create-gh-relase-action and upload-rust-binary-action to abstract
the release builds. This also makes it easier to cross compile for
different architectures using cross and cargo-zigbuild.

cargo-zigbuild also makes it possible to request a specific glibc version
to compile against, so we can bump the runner version without worrying
about binary compatibility.

This tries to use native runners where possible. When we get issues with
quota we should move most jobs to x86 and cross compile more.

The logic for this is mostly taken from cargo-nextest's
[release ci](https://github.com/nextest-rs/nextest/blob/8832117c2700a07db6870e0087724f1b40e4292a/.github/workflows/release.yml).

Closes: https://github.com/neocmakelsp/neocmakelsp/issues/203